Revert "Refactor code folder structure"
