+++ py_script/brushMenuDialog.py
@@ -1,9 +1,10 @@
+import sys
 
-from PyQt5 import uic
+from PyQt5 import uic, QtCore
 from PyQt5.QtWidgets import *
-from utils.utils import *
+from utils import *
 
-brushMenu_ui = '../../ui_design/brushMenuDialog.ui'
+brushMenu_ui = '../ui_design/brushMenuDialog.ui'
 
 form_brushMenu = resource_path(brushMenu_ui)
 form_class_brushMenu = uic.loadUiType(form_brushMenu)[0]
@@ -22,17 +23,13 @@ class BrushMenu(QDialog, form_class_brushMenu):
         #self.lineEdit.setText(f'{self.brushSize} px') 
 
     def changeSliderValueText(self):
-        number = self.horizontalSlider.value()
-
-        if number % 2 == 1: 
-            number += 1
-        self.brushSize = number
-        self.lineEdit.setText(f'{number} px')
-        self.horizontalSlider.setValue(number) 
-            
+        self.brushSize = self.horizontalSlider.value()
+        self.lineEdit.setText(f'{self.brushSize} px')
 
     def keyPressEvent(self, event):
-        if event.key() in [16777220, 16777221] : # enter key code
+        # print(event.key())
+        if event.key() in [16777220, 16777221] : 
+            # print('Enter pressed')
             self.changeBrushSizeAndSliderBar()
         else:
             super().keyPressEvent(event)
@@ -40,17 +37,8 @@ class BrushMenu(QDialog, form_class_brushMenu):
     def changeBrushSizeAndSliderBar(self):
         txt = self.lineEdit.text()
         numbers = [int(s) for s in txt.split() if s.isdigit()]
-        number = numbers[0]
-        if number : 
-            if number < 2: 
-                number = 2
-            elif number > 50: # change it to max number 
-                number = 50
-
-            if (number % 2) == 1:
-                number += 1
-
-            self.brushSize = number
+        if numbers : 
+            self.brushSize = numbers[0]
             self.lineEdit.setText(f'{self.brushSize} px') 
             self.horizontalSlider.setValue(self.brushSize)
